Revert "Merge branch 'fix_is_pyramid' of https://github.com/NHPatterson/napari into NHPatterson-fix_is_pyramid"
This reverts commit 285ae8407672debb619edc905a055b0dda7d90d5, reversing
changes made to bc09f23e1a6b8f744b4615009b000aeaca3f916a.

+++ napari/util/misc.py
@@ -65,7 +65,7 @@ def is_rgb(shape):
 def is_pyramid(data):
     """If shape of arrays along first axis is strictly decreasing.
     """
-    size = np.array([np.prod(d.shape, dtype=np.uint64) for d in data])
+    size = np.array([np.prod(d.shape) for d in data])
     if len(size) > 1:
         return np.all(size[:-1] > size[1:])
     else:
